+++ packages/e2e-test-utils/src/test/get-page-error.js
@@ -1,9 +1,9 @@
 /**
  * Internal dependencies
  */
-import { hasPageError } from '../has-page-error';
+import { getPageError } from '../get-page-error';
 
-describe( 'hasPageError', () => {
+describe( 'getPageError', () => {
 	let originalPage;
 
 	beforeEach( () => {
@@ -14,31 +14,31 @@ describe( 'hasPageError', () => {
 		global.page = originalPage;
 	} );
 
-	it( 'returns false if there is no error', async () => {
+	it( 'resolves to null if there is no error', async () => {
 		global.page = {
 			content: () => 'Happy!',
 		};
 
-		expect( await hasPageError() ).toBe( false );
+		expect( await getPageError() ).toBe( null );
 	} );
 
 	it.each( [
 		[
 			'PHP, HTML',
-			'<b>Notice</b>:  Undefined property: WP_Block_Type_Registry::$x in <b>/var/www/html/wp-content/plugins/gutenberg/lib/blocks.php</b> on line <b>47</b><br />',
+			'<b>Notice</b>:  Undefined property: WP_Block_Type_Registry::$x in <b>/var/www/html/wp-content/plugins/gutenberg/lib/blocks.php</b> on line <b>47</b>',
 		],
 		[
 			'PHP, Plaintext',
 			'Notice: Undefined property: WP_Block_Type_Registry::$x in /var/www/html/wp-content/plugins/gutenberg/lib/blocks.php on line 47',
 		],
 	] )(
-		'returns true if there is an error (%s)',
+		'resolves to the error message if there is an error (%s)',
 		async ( _variant, error ) => {
 			global.page = {
 				content: () => error,
 			};
 
-			expect( await hasPageError() ).toBe( true );
+			expect( await getPageError() ).toBe( error );
 		}
 	);
 } );